wip: backup before baseline reset
