+++ client/src/lib/api.js
@@ -1,0 +1,1 @@
+export const API = import.meta.env.VITE_API_BASE || "http://localhost:8080";
